Fixes issue #11 by considering advanced and important assignments (#16)
* Fixes issue #11 by considering advanced and important assignments

Advanced and important assignments were not considered on the back end.
To add these with ease I cre static to_string to TypeOfCommittee. Also made some formating changes in the user_newTeacherPage()
process for us. Had to add a to_string to typeOfCommittee.

* Fixed two bugs on user_newTeacherPage and added .gitignore file

Fixed the error messages sent to the user when they try to get
assignments that are not available. Also found a bug with 6th grade
available numbers not showing up correctly on the page.

Added a .gitignore file to remove the db file from tracking.

* Partial assignments are possible

When a users asks for more assignments of a certain type then possible,
the system will try to give the user as many of those assignments as
possible and then return an warning message.

* First warning flash out of loop

Moved the first warning flash out of the for loop for it to only be sent
once. Also changed the wording to make more sense.